==> Research/01-Data_Ingestion.ipynb <==
import os

%pwd

os.chdir("../")

%pwd

from ml_Project import logger

from urllib import request

dir = "Dataset/Data_Ingestion"
file_path = "Dataset/Data_Ingestion/data.zip"
url = "https://github.com/RawatRahul14/diabetes_dataset/raw/main/diabetes_dataset.zip"

from ml_Project.utils.common import create_directories

def create_dir(dir):
    create_directories([dir])

create_dir(dir)

def download_file(file_path, url):
    if not os.path.exists(file_path):
        directory = os.path.dirname(file_path)
        filename, headers = request.urlretrieve(
            url = url,
            filename = file_path
        )
        logger.info(f"{filename} download! with the following info: \n{headers}")
    else:
        logger.info(f"File already exists.")

import zipfile

def extract_zip_file(unzip_filepath, file_path):
    unzip_file_path = unzip_filepath
    os.makedirs(unzip_file_path, exist_ok = True)
    with zipfile.ZipFile(file_path, "r") as zip_ref:
        zip_ref.extractall(unzip_file_path)

download_file(file_path, url)

unzip_filepath = "Dataset/Data_Ingestion"

extract_zip_file(unzip_filepath, file_path)

from ml_Project import logger
from urllib import request
import zipfile

class DataIngestion:
    def __init__(self, data_dir, file_path, url, unzip_filepath):
        self.data_dir = data_dir
        self.file_path = file_path
        self.url = url
        self.unzip_filepath = unzip_filepath

    def create_dir(self):
        create_directories([self.data_dir])

    def download_file(self):
        if not os.path.exists(self.file_path):
            filename, headers = request.urlretrieve(
                url = self.url,
                filename = self.file_path
            )
            logger.info(f"{filename} download! with the following info: \n{headers}")
        else:
            logger.info(f"File already exists.")

    def extract_zip_file(self):
        unzip_file_path = self.unzip_filepath
        os.makedirs(unzip_file_path, exist_ok = True)
        with zipfile.ZipFile(self.file_path, "r") as zip_ref:
            zip_ref.extractall(unzip_file_path)

data_dir = "Dataset/Data_Ingestion"
file_path = "Dataset/Data_Ingestion/data.zip"
url = "https://github.com/RawatRahul14/diabetes_dataset/raw/main/diabetes_dataset.zip"
unzip_filepath = "Dataset/Data_Ingestion"

try:
    data_ingestion = DataIngestion(data_dir = data_dir,
                                   file_path = file_path,
                                   url = url,
                                   unzip_filepath = unzip_filepath)
    
    data_ingestion.create_dir()
    data_ingestion.download_file()
    data_ingestion.extract_zip_file()
except Exception as e:
    raise e
